requirements_full.txt
streamlit
pandas
matplotlib
fpdf2
pillow
opencv-python
numpy
ultralytics
openai
python-dotenv
torch
torchvision
sentence-transformers
transformers
scikit-learn
tqdm
jinja2